+++ kakutou/pages/playgame.js
@@ -1,12 +1,13 @@
 import {DataContext} from '../DataContext'
-import {useContext, useEffect} from 'react'
+import {useContext, useEffect, useState} from 'react'
 
 
 
 export default function Playgame() {
+    const [msg,setMsg]=useState(null)
     const context = useContext(DataContext)
     const socket = context.socket
-    const SocketOn = ()=> {socket.on('test',()=>console.log("test.ok"))}
+    const SocketOn = ()=> {socket.on('test',(msg)=>{console.log("test.ok");setMsg(msg)})}
 
     useEffect(()=> {SocketOn()
     return()=>{socket.off("test")}})
@@ -15,6 +16,7 @@ export default function Playgame() {
 return(
     <>
     <h2>hello</h2>
+    {msg}
     </>
 )
 }
